Update theme_icon_font.sh
Updated theme_icon_font.sh

+++ theme_icon_font.sh
@@ -1,19 +1,27 @@
+# check if font file is downloaded to ~/Downloads/
+if [ ! -f ~/Downloads/SF-Pro.dmg ]
+then
+	echo "ERROR: SF-Pro.dmg is not downloaded to the Downloads folder. Download it before proceeding!"
+	exit 1
+fi
+
 cd ~/
 mkdir .themes
 mkdir .icons
 mkdir .fonts
+
 # theme
 cd ~/Downloads/
 git clone https://github.com/vinceliuice/WhiteSur-gtk-theme.git
-#cd ~/Downloads/WhiteSur-gtk-theme/release/
 tar -xvf ~/Downloads/WhiteSur-gtk-theme/release/WhiteSur-Dark.tar.xz
 mv -v ~/Downloads/WhiteSur-Dark/ ~/.themes
+
 # icon
 cd ~/Downloads/
 git clone https://github.com/PapirusDevelopmentTeam/papirus-icon-theme.git
 mv -v ~/Downloads/papirus-icon-theme/Papirus/ ~/.icons
+
 # font
-# font file has to be downloaded
 cd ~/Downloads/
 7z x SF-Pro.dmg
 cd SFProFonts/
@@ -22,6 +30,7 @@ cd SFProFonts/
 cd Library
 mv Fonts/ SF\ Pro/ # rename
 mv "SF Pro" ~/.fonts
+
 # delete resources
 cd ~/Downloads/
 rm -fv SF-Pro.dmg
